backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
pandas
transformers
sentence-transformers
torch
scipy
scikit-learn
grpcio
grpcio-tools
googleapis-common-protos
google
confluent-kafka
redis